+++ xbar/XbarMflitRTL.py
@@ -1,11 +1,11 @@
 '''
 ==========================================================================
-MeshRouterMFlitRTL
+XbarMflitRTL.py
 ==========================================================================
-Mesh router that supports multi-flit packet.
+A crossbar that supports multi-flit packet.
 
 Author : Yanghui Ou
-  Date : Feb 13, 2020
+  Date : Feb 18, 2020
 '''
 from pymtl3 import *
 from pymtl3.stdlib.ifcs import RecvIfcRTL, SendIfcRTL
@@ -13,14 +13,15 @@ from ocnlib.utils import get_nbits
 from router.InputUnitRTL import InputUnitRTL
 from router.OutputUnitRTL import OutputUnitRTL
 from router.SwitchUnitGrantHoldRTL import SwitchUnitGrantHoldRTL
+from router.SwitchUnitNullRTL import SwitchUnitNullRTL
 
-from .MeshRouteUnitRTLMFlitXY import MeshRouteUnitRTLMFlitXY
+from .XbarRouteUnitMflitRTL import XbarRouteUnitMflitRTL
 
 #-------------------------------------------------------------------------
-# MeshRouterMFlitRTL
+# XbarMflitRTL
 #-------------------------------------------------------------------------
 
-class MeshRouterMFlitRTL( Component ):
+class XbarMflitRTL( Component ):
 
   #-----------------------------------------------------------------------
   # Construct
@@ -28,31 +29,34 @@ class MeshRouterMFlitRTL( Component ):
 
   def construct( s,
     Header,
-    PositionType,
+    num_inports    = 2,
+    num_outports   = 2,
     InputUnitType  = InputUnitRTL,
-    RouteUnitType  = MeshRouteUnitRTLMFlitXY,
+    RouteUnitType  = XbarRouteUnitMflitRTL,
     SwitchUnitType = SwitchUnitGrantHoldRTL,
     OutputUnitType = OutputUnitRTL,
   ):
 
     # Local parameter
 
-    s.num_inports  = 5
-    s.num_outports = 5
+    s.num_inports  = num_inports
+    s.num_outports = num_outports
     s.PhitType     = mk_bits( get_nbits( Header ) )
+
+    # Special case for num_inports = 1
+    if num_inports == 1: SwitchUnitType = SwitchUnitNullRTL
 
     # Interface
 
     s.recv = [ RecvIfcRTL( s.PhitType ) for _ in range( s.num_inports  ) ]
     s.send = [ SendIfcRTL( s.PhitType ) for _ in range( s.num_outports ) ]
-    s.pos  = InPort( PositionType )
 
     # Components
 
-    s.input_units  = [ InputUnitType( s.PhitType )
-                       for _ in range( s.num_inports ) ]
+    s.input_units = [ InputUnitType( s.PhitType )
+                      for _ in range( s.num_inports )  ]
 
-    s.route_units  = [ RouteUnitType( Header, PositionType )
+    s.route_units  = [ RouteUnitType( Header, s.num_outports )
                        for i in range( s.num_inports ) ]
 
     s.switch_units = [ SwitchUnitType( s.PhitType, s.num_inports )
@@ -66,7 +70,6 @@ class MeshRouterMFlitRTL( Component ):
     for i in range( s.num_inports ):
       s.recv[i]             //= s.input_units[i].recv
       s.input_units[i].give //= s.route_units[i].get
-      s.pos                 //= s.route_units[i].pos
 
     for i in range( s.num_inports ):
       for j in range( s.num_outports ):
@@ -84,4 +87,4 @@ class MeshRouterMFlitRTL( Component ):
   def line_trace( s ):
     in_trace  = '|'.join([ f'{ifc}' for ifc in s.recv ])
     out_trace = '|'.join([ f'{ifc}' for ifc in s.send ])
-    return f'{in_trace}({s.pos}){out_trace}'
+    return f'{in_trace}(){out_trace}'
