+++ xbar/XbarRouteUnitMflitRTL.py
@@ -1,8 +1,8 @@
 '''
 ==========================================================================
-MeshRouteUnitRTLMFlitXY.py
+XbarRouteUnitMflitRTL.py
 ==========================================================================
-Route unit for mesh that uses XY-routing and supports multi-flit packet.
+Route unit for crossbar that supports multi-flit packet.
 
 - NOTE:
 
@@ -10,26 +10,27 @@ It requires no tail flit and that the header flit should have a field that
 indicates the length of the payload in terms of number of flits.
 
 Authour : Yanghui Ou
-   Date : Feb 9, 2020  
+   Date : Feb 18, 2020
 '''
 from pymtl3 import *
 from pymtl3.stdlib.ifcs import GetIfcRTL, GiveIfcRTL
 from ocnlib.rtl import Counter, GrantHoldArbiter
 from ocnlib.utils import get_nbits, get_plen_type
 from ocnlib.utils.connects import connect_bitstruct
-from .directions import *
 
-class MeshRouteUnitRTLMFlitXY( Component ):
+class XbarRouteUnitMflitRTL( Component ):
 
   #-----------------------------------------------------------------------
   # construct
   #-----------------------------------------------------------------------
 
-  def construct( s, HeaderFormat, PositionType, plen_field_name='plen' ):
+  def construct( s, HeaderFormat, num_outports=4, plen_field_name='plen' ):
     # Meta data
-    s.num_outports = 5
+    s.num_outports = num_outports
     s.HeaderFormat = HeaderFormat
     s.PhitType     = mk_bits( get_nbits( HeaderFormat ) )
+    s.dir_nbits    = clog2( num_outports ) if num_outports > 1 else 1
+    s.DirType      = mk_bits( s.dir_nbits )
     s.STATE_HEADER = b1(0)
     s.STATE_BODY   = b1(1)
 
@@ -37,7 +38,6 @@ class MeshRouteUnitRTLMFlitXY( Component ):
 
     # Interface
     s.get  = GetIfcRTL( s.PhitType )
-    s.pos  = InPort( PositionType ) # TODO: figure out a way to encode position
 
     s.give = [ GiveIfcRTL( s.PhitType ) for _ in range( s.num_outports ) ]
     s.hold = [ OutPort( Bits1 ) for _ in range( s.num_outports ) ]
@@ -46,8 +46,8 @@ class MeshRouteUnitRTLMFlitXY( Component ):
     s.header      = Wire( HeaderFormat )
     s.state       = Wire( Bits1 )
     s.state_next  = Wire( Bits1 )
-    s.out_dir_r   = Wire( Bits3 )
-    s.out_dir     = Wire( Bits3 )
+    s.out_dir_r   = Wire( s.DirType )
+    s.out_dir     = Wire( s.DirType )
     s.any_give_en = Wire( Bits1 )
 
     s.counter = Counter( PLenType )(
@@ -56,8 +56,8 @@ class MeshRouteUnitRTLMFlitXY( Component ):
     )
 
     connect_bitstruct( s.get.ret, s.header )
-    
-    for i in range( 5 ):
+
+    for i in range( s.num_outports ):
       s.get.ret //= s.give[i].ret
     s.get.en //= s.any_give_en
 
@@ -111,17 +111,8 @@ class MeshRouteUnitRTLMFlitXY( Component ):
     @s.update
     def up_out_dir():
       if ( s.state == s.STATE_HEADER ) & s.get.rdy:
-        if ( s.header.dst_x == s.pos.pos_x ) & ( s.header.dst_y == s.pos.pos_y ):
-          s.out_dir = b3( SELF )
-        elif s.header.dst_x < s.pos.pos_x:
-          s.out_dir = b3( WEST )
-        elif s.header.dst_x > s.pos.pos_x:
-          s.out_dir = b3( EAST )
-        elif s.header.dst_y < s.pos.pos_y:
-          s.out_dir = b3( SOUTH )
-        else:
-          s.out_dir = b3( NORTH )
-       
+        s.out_dir = s.header.dst[0:s.dir_nbits]
+
       else:
         s.out_dir = s.out_dir_r
 
@@ -132,8 +123,8 @@ class MeshRouteUnitRTLMFlitXY( Component ):
     @s.update
     def up_give_rdy_hold():
       for i in range( s.num_outports ):
-        s.give[i].rdy = ( b3(i) == s.out_dir ) & s.get.rdy
-        s.hold[i]     = ( b3(i) == s.out_dir ) & ( s.state == s.STATE_BODY )
+        s.give[i].rdy = ( s.DirType(i) == s.out_dir ) & s.get.rdy
+        s.hold[i]     = ( s.DirType(i) == s.out_dir ) & ( s.state == s.STATE_BODY )
 
   #-----------------------------------------------------------------------
   # line_trace
@@ -142,15 +133,9 @@ class MeshRouteUnitRTLMFlitXY( Component ):
   def line_trace( s ):
     give_trace = '|'.join([ f'{ifc}' for ifc in s.give ])
     hold  = ''.join([ '^' if h else '.' for h in s.hold ])
-    pos   = f'<{s.pos.pos_x},{s.pos.pos_y}>'
     count = f'{s.counter.count}'
     state = 'H' if s.state == s.STATE_HEADER else \
             'B' if s.state == s.STATE_BODY else   \
             '!'
-    dir   = 'n' if s.out_dir == NORTH else \
-            's' if s.out_dir == SOUTH else \
-            'w' if s.out_dir == WEST  else \
-            'e' if s.out_dir == EAST  else \
-            'S' if s.out_dir == SELF  else \
-            '?'
-    return f'{s.get}({pos}[{state}{count}]{dir}{hold}){give_trace}'
+    dir   = f'{s.out_dir}'
+    return f'{s.get}([{state}{count}]{dir}{hold}){give_trace}'
